Fixed a warning (strlen returned size_t ??, %d expects int)
git-svn-id: https://gamescrafters.svn.sourceforge.net/svnroot/gamescrafters/GamesmanClassic/trunk@2845 a668be5d-ef11-cb4c-8bfd-e08cbb24f2ba

+++ gamesman/src/mothello.c
@@ -950,7 +950,7 @@ BOOLEAN ValidTextInput (STRING input)
 {
 	char alphabet[] = "ABCDEFGHIJKLMNOPQRSTUVWXYZ";
 	
-	if(DEBUG) printf("ValidTextInput Starting. Strlen = %d\n", strlen(input));
+	if(DEBUG) printf("ValidTextInput Starting. Strlen = %d\n", (int) strlen(input));
 	
 	if(strlen(input) == 1)
 	{
